Renaming actionTypes to index so we can just import the constants folder
