+++ app/reducers/topic.js
@@ -3,7 +3,7 @@ import {
   CREATE_TOPIC,
   DESTROY_TOPIC,
   INCREMENT_COUNT,
-  DECREMENT_COUNT } from 'constants/actionTypes';
+  DECREMENT_COUNT } from 'constants';
 
 
 export default function topic(state = {
